class 7

+++ Tutorial/src/com/tutorial/main/FastEnemy.java
@@ -9,7 +9,7 @@ public class FastEnemy extends GameObject {
 	private Handler handler;
 	
 	
-	public FastEnemy(int x, int y, ID id, Handler handler) {
+	public FastEnemy(float x, float y, ID id, Handler handler) {
 		super(x, y, id);
 		this.handler = handler;
 		
@@ -37,12 +37,12 @@ public class FastEnemy extends GameObject {
 	@Override
 	public void render(Graphics g) {
 		g.setColor(Color.magenta);
-		g.fillRect(x, y, 16, 16);
+		g.fillRect((int)x, (int)y, 16, 16);
 	}
 
 	@Override
 	public Rectangle getBounds() {
-		return new Rectangle(x,y,16,16);
+		return new Rectangle((int)x,(int)y,16,16);
 	}
