+++ Tutorial/src/com/tutorial/main/HUD.java
@@ -32,8 +32,8 @@ public class HUD {
 	
 	public void tick() {
 		
-		HEALTH = Game.clamp(HEALTH, 0, 100);
-		greenValue = Game.clamp(greenValue, 0, 255);
+		HEALTH = (int) Game.clamp(HEALTH, 0, 100);
+		greenValue =(int) Game.clamp(greenValue, 0, 255);
 		
 		//changes greenValue for the life bar
 		greenValue = HEALTH * 2;
